requirements.txt
pytesseract
pdf2image
pillow
sumy
pandas
tk
python-docx
opencv-python
